+++ src/main/java/frc/robot/RobotContainer.java
@@ -10,6 +10,7 @@ import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.commands.ExampleCommand;
 import frc.robot.subsystems.ExampleSubsystem;
+import frc.utils.laptopcomms.RobotReciever;
 
 /**
  * This class is where the bulk of the robot should be declared. Since Command-based is a
@@ -20,6 +21,7 @@ import frc.robot.subsystems.ExampleSubsystem;
 public class RobotContainer {
 
     private final ExampleSubsystem m_exampleSubsystem = new ExampleSubsystem();
+    private final RobotReciever ROBOT_RECIEVER = new RobotReciever();
 
     /**
      * The container for the robot. Contains subsystems, OI devices, and commands.
